+++ example/lib/screens/edit_alarm.dart
@@ -97,7 +97,6 @@ class _ExampleAlarmEditScreenState extends State<ExampleAlarmEditScreen> {
       notificationTitle: showNotification ? 'Alarm example' : null,
       notificationBody: showNotification ? 'Your alarm ($id) is ringing' : null,
       assetAudioPath: assetAudio,
-      stopOnNotificationOpen: false,
     );
     return alarmSettings;
   }